refactor reorg
